+++ lib/s_universal_html/src/web_actions_impl_others.dart
@@ -1,0 +1,7 @@
+void reloadWindow() {}
+
+void navigateTo(String url) {}
+
+void replaceLocation(String url) {}
+
+String get currentHref => '';